refactor : IDE-121 파일트리 util 함수들 기능에 따른 분리
